세번째 코밋

+++ untact/src/main/java/com/sbs/untact/controller/UsrArticleController.java
@@ -7,6 +7,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -31,16 +33,25 @@ public class UsrArticleController {
 	
 	@RequestMapping("usr/article/detail")
 	@ResponseBody
-	public Article showDetail(int id) {
+	public ResultData showDetail(Integer id) {
+		if( id == null ) {
+			return new ResultData("f-1", "id를 입력해주세요");
+		}
 		
-		Article article = articleService.getArticle(id);
+		Article article = articleService.getForPrintArticle(id);
 		
-		return article;
+		if (article == null) {
+			return new ResultData("f-2", "존재하지 않는 게시물 번호입니다.");
+		}
+		
+		return new ResultData("s-1","성공","article",article);
 	}
 	
 	@RequestMapping("usr/article/list")
 	@ResponseBody
-	public List<Article> showList( String searchKeywordType, String searchKeyword){
+	public ResultData showList( String searchKeywordType, String searchKeyword,
+								@RequestParam(defaultValue="1") int page){
+		
 		System.out.println("searchKeyword ============== "+searchKeyword);
 		System.out.println("searchKeywordType ============== "+searchKeywordType);
 		if(searchKeywordType != null) {
@@ -59,13 +70,30 @@ public class UsrArticleController {
 			searchKeyword = searchKeyword.trim();
 		}
 		
-		return articleService.getArticles(searchKeywordType, searchKeyword);
+		if( searchKeyword == null ) {
+			searchKeywordType = null;
+		}
 		
+		// 한페이지에 아이템을 몇개를 보여줄것인가.
+		int itemsInAPage = 20;
+		
+		List<Article> articles = articleService.getForPrintArticles(searchKeywordType, searchKeyword, page ,itemsInAPage);
+		
+		return new ResultData("s-1","성공","articles",articles);
 	}
 	
 	@RequestMapping("usr/article/doAdd")
 	@ResponseBody
-	public ResultData doAdd(@RequestParam Map<String, Object> param){
+	public ResultData doAdd(@RequestParam Map<String, Object> param, HttpSession session){
+		
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+		
+		// getAsInt 메서드를 탓을때 첫번째 매개변수에 값이 있다면 int로 반환되서 
+		// loginedMemberId 변수에 담길것이고 값이 없다면 defaultValue 로 지정한
+		// 두번쨰 매개변수 자리의 값이 그대로 출력될것이다, 이것을 이용하여 메서드를 구현한다.
+//		if(loginedMemberId == 0) {
+//			return new ResultData("f-2", "로그인후 이용해주세요");
+//		}
 		
 		if (param.get("title") == null) {
 			return new ResultData("f-1","제목을 입력해주세여." );
@@ -74,6 +102,9 @@ public class UsrArticleController {
 		if (param.get("content") == null) {
 			return new ResultData("f-1","내용을 입력해주세요." );
 		}
+		
+		// param에 memberId 를 키값으로 loginedMemberId 를 담아 보낼것이다.
+		param.put("memberId", loginedMemberId);
 
 		ResultData rsData = articleService.addArticle(param);
 		
@@ -83,7 +114,11 @@ public class UsrArticleController {
 	
 	@RequestMapping("usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(Integer id){
+	public ResultData doDelete(Integer id, HttpSession session){
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId") , 0);
+//		if(loginedMemberId == 0 ) {
+//			return new ResultData("f-2","로그인후 이용해주세요");
+//		}
 		
 		if(id == null) {
 			return new ResultData("f-1","id를 입력해주세요");
@@ -95,6 +130,12 @@ public class UsrArticleController {
 			return new ResultData ("f-1", "해당 게시글은 존재하지 않습니다." );
 		}
 		
+		ResultData actorCanDeleteRd = articleService.getActorCanDeleteRd(article, loginedMemberId);
+		
+		if(actorCanDeleteRd.isFail()) {
+			return actorCanDeleteRd;
+		}
+		
 		return articleService.deleteArticle(id);
 		
 	}
@@ -102,8 +143,13 @@ public class UsrArticleController {
 	
 	@RequestMapping("usr/article/doModify")
 	@ResponseBody
-	public ResultData doModify(Integer id, String title, String content){
+	public ResultData doModify(Integer id, String title, String content, HttpSession session){
 		
+		// doModify 메서드 역시 session 인증을 거쳐야 함으로 아래 로직으로 세션에 id가 있는지 체크부터함
+		int loginedMemberId = Util.getAsInt(session.getAttribute("loginedMemberId"), 0);
+//		if( loginedMemberId == 0) {
+//			return new ResultData("f-2", "로그인후 이용해 주세요");
+//		}
 		
 		if(id == null) {
 			return new ResultData("f-1","id를 입력해주세요");
@@ -121,9 +167,16 @@ public class UsrArticleController {
 			return new ResultData("f-1","해당 게시물은 존재하지않습ㄴ디ㅏ.","id",id);
 		}
 		
+		ResultData actorCanModifyRd = articleService.getActorCanModifyRd(article, loginedMemberId);
+		
+		if(actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
+		
 		return articleService.modifyArticle(id,title,content);
 			
 	}
 	
 	
 }
+
